double dT

+++ src/FusionEKF.cpp
@@ -149,7 +149,7 @@ void FusionEKF::ProcessMeasurement(const MeasurementPackage &measurement_pack) {
        * Update the process noise covariance matrix.
        * Use noise_ax = 9 and noise_ay = 9 for your Q matrix.
      */
-    float dT = (measurement_pack.timestamp_ - previous_timestamp_)/ 1000000.0;
+    double dT = (measurement_pack.timestamp_ - previous_timestamp_)/ 1000000.0;
 
     ekf_.F_(0, 2) = dT;
     ekf_.F_(1, 3) = dT;
